requirements.txt
langchain
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
streamlit
python-dotenv
huggingface_hub
transformers